lib: remove the 'daemon' module to reduce depth
