+++ src/bitcoind/utils.rs
@@ -1,4 +1,4 @@
-use crate::daemon::{
+use crate::{
     bitcoind::{interface::UtxoInfo, BitcoindError},
     database::{
         interface::{
